% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_GISAID.R
\name{build_GISAID}
\alias{build_GISAID}
\title{Build GISAID reference information}
\usage{
build_GISAID(file, type, seq_type, nROW)
}
\arguments{
\item{file}{the address of the fasta file.}

\item{type}{flu type}

\item{seq_type}{DNA or AA.}

\item{nROW}{rows in df}
}
\description{
return a list contain meta information of virus from a fasta file.
}
