% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_Serum_new.R
\name{build_Serum_new}
\alias{build_Serum_new}
\title{title}
\usage{
build_Serum_new(TB_list, GISAID_ref, type)
}
\arguments{
\item{TB_list}{description}

\item{GISAID_ref}{description}

\item{type}{description}
}
\description{
d
}
