% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_matched.R
\name{check_matched}
\alias{check_matched}
\title{Check row matched}
\usage{
check_matched(TB_list)
}
\arguments{
\item{TB_list}{description}
}
\description{
check which row in a table with Genetic group do not
match to a GISAID sequence.
}
