% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CN2EN.R
\name{CN2EN}
\alias{CN2EN}
\title{Turn Chinese virus names to English}
\usage{
CN2EN(TB_list, parallel = TRUE)
}
\arguments{
\item{TB_list}{a table list.}

\item{parallel}{a logical value to indicate whether to use parallel processing. Default is TRUE.}
}
\description{
Use pinyin package to translate Chinese Virus names to English to match GISAID sequences.
}
