% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exact_matching_strain.R
\name{exact_matching_strain}
\alias{exact_matching_strain}
\title{exact match strain from GISAID}
\usage{
exact_matching_strain(absent_Isolates, GISAID_REF, save_path)
}
\arguments{
\item{absent_Isolates}{description}

\item{GISAID_REF}{description}

\item{save_path}{description}
}
\description{
在GISAID参考信息中,寻找Virus name, Collection date, Passage history完全一样的序列,并匹配EPI_ISL_ID并写入文件
}
