% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_Isolate_ID_absent.R
\name{extract_Isolate_ID_absent}
\alias{extract_Isolate_ID_absent}
\title{extract}
\usage{
extract_Isolate_ID_absent(TB_list, absent.row, info_col, only_Ref)
}
\arguments{
\item{TB_list}{description}

\item{absent.row}{dd}

\item{info_col}{description}

\item{only_Ref}{dd}
}
\description{
A short description...
}
