% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GISAID_standardization.R
\name{GISAID_standardization}
\alias{GISAID_standardization}
\title{title}
\usage{
GISAID_standardization(names, type = "H3N2")
}
\arguments{
\item{names}{description}

\item{type}{description}
}
\description{
A short description...
}
