% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_dates.R
\name{match_dates}
\alias{match_dates}
\title{Obtain index of GISAID}
\usage{
match_dates(Dates, GISAID_ref)
}
\arguments{
\item{Dates}{description}

\item{GISAID_ref}{description}
}
\description{
A short description...
}
