% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_Names.R
\name{match_Names}
\alias{match_Names}
\title{Obtain index of GISAID}
\usage{
match_Names(tb, GISAID_ref, type = "H3N2")
}
\arguments{
\item{GISAID_ref}{description}

\item{type}{description}

\item{Names}{description}
}
\description{
A short description...
}
