% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_passage.R
\name{match_passage}
\alias{match_passage}
\title{Obtain index of GISAID}
\usage{
match_passage(Passages, GISAID_ref)
}
\arguments{
\item{Passages}{description}

\item{GISAID_ref}{description}
}
\description{
A short description...
}
