% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchGISAIDApp.R
\name{matchGISAIDApp}
\alias{matchGISAIDApp}
\title{interactive GISAID match and fill，return data.frame}
\usage{
matchGISAIDApp(strain_df, GISAID_df)
}
\arguments{
\item{strain_df}{data.frame，must have Virus column}

\item{GISAID_df}{data.frame，must have Isolate name、Isolate ID、Passage details/history columns}
}
\value{
modified strain_df（data.frame）
}
\description{
interactive GISAID match and fill，return data.frame
}
