% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Names_matched_idx.R
\name{Names_matched_idx}
\alias{Names_matched_idx}
\title{Obtain index of GISAID}
\usage{
Names_matched_idx(V_names, GISAID_ref)
}
\arguments{
\item{V_names}{description}

\item{GISAID_ref}{description}
}
\description{
A short description...
}
