% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sequences.R
\name{read_sequences}
\alias{read_sequences}
\title{title}
\usage{
read_sequences(file, seq_type)
}
\arguments{
\item{file}{description}

\item{seq_type}{description}
}
\description{
A short description...
}
