% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_record_without_EPI_ID.R
\name{rm_record_without_EPI_ID}
\alias{rm_record_without_EPI_ID}
\title{Build GISAID reference information}
\usage{
rm_record_without_EPI_ID(TB_list)
}
\arguments{
\item{TB_list}{description}
}
\description{
return a list contain meta information of virus from a fasta file.
}
