% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TB_list_unfolded.R
\name{TB_list_unfolded}
\alias{TB_list_unfolded}
\title{dd}
\usage{
TB_list_unfolded(
  TB_list,
  GISAID_ref,
  type,
  parallel = TRUE,
  match.col,
  match_start
)
}
\description{
A short description...
}
