% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WHO_standardization.R
\name{WHO_standardization}
\alias{WHO_standardization}
\title{title}
\usage{
WHO_standardization(names, type = "H3N2")
}
\arguments{
\item{names}{description}

\item{type}{description}
}
\description{
A short description...
}
